+++ src/plugin.spec.ts
@@ -1,11 +1,11 @@
 import { existsSync, rmSync } from 'fs'
 import path from 'path'
 import { expect } from 'vitest'
-import { compareSnapshots, updateSnapshot, type CompareSnapshotOptions, type UpdateSnapshotOptions } from './plugin'
+import { compareSnapshots, updateSnapshot, type VisualRegressionOptions, type UpdateSnapshotOptions } from './plugin'
 
 const baseUpdateOptions: UpdateSnapshotOptions = {
   screenshotName: 'enjuto',
-  specName: 'sub-folder',
+  type: 'regression',
   screenshotAbsolutePath: path.join('./src/fixtures/assets/base/enjuto.png'),
   baseDirectory: path.join('./src/fixtures/assets/base'),
   spec: {
@@ -15,9 +15,13 @@ const baseUpdateOptions: UpdateSnapshotOptions = {
   }
 }
 
-const baseCompareOptions: CompareSnapshotOptions = {
+const baseCompareOptions: VisualRegressionOptions = {
   ...baseUpdateOptions,
-  errorThreshold: 0.1,
+  pluginOptions: {
+    errorThreshold: 0.1,
+    failSilently: false,
+    pixelmatchOptions: {}
+  },
   screenshotOptions: {},
   diffDirectory: path.join('./src/fixtures/diff'),
   generateDiff: 'fail'
@@ -29,7 +33,7 @@ describe('plugin', () => {
       it('should copy the snapshot to default folder', async () => {
         const options: UpdateSnapshotOptions = {
           ...baseUpdateOptions,
-          baseDirectory: undefined,
+          baseDirectory: 'cypress/snapshots/base',
           screenshotName: 'defaultName'
         }
         const result = await updateSnapshot(options)
@@ -90,11 +94,14 @@ describe('plugin', () => {
       })
       describe('when image differs', () => {
         it('should generate a diff image by default', async () => {
-          const options: CompareSnapshotOptions = {
+          const options: VisualRegressionOptions = {
             ...baseCompareOptions,
             screenshotAbsolutePath: path.join('./src/fixtures/assets/mod/random.png'),
-            errorThreshold: 0,
-            specName: '',
+            pluginOptions: {
+              errorThreshold: 0,
+              failSilently: false,
+              pixelmatchOptions: {}
+            },
             screenshotName: 'random'
           }
 
@@ -106,15 +113,18 @@ describe('plugin', () => {
           expect(result.images.diff).toBeDefined()
           expect(result.baseGenerated).toBeUndefined()
           expect(result.error).toBeDefined()
-          const isDiffGenerated = existsSync(path.join(options.diffDirectory as string, 'random.png'))
+          const isDiffGenerated = existsSync(path.join(options.diffDirectory, 'random.png'))
           expect(isDiffGenerated).toBe(true)
         })
         it('should not generate a diff image if generateDiff is set to never', async () => {
-          const options: CompareSnapshotOptions = {
+          const options: VisualRegressionOptions = {
             ...baseCompareOptions,
             screenshotAbsolutePath: path.join('./src/fixtures/assets/mod/not_ever.png'),
-            errorThreshold: 0,
-            specName: '',
+            pluginOptions: {
+              errorThreshold: 0,
+              failSilently: false,
+              pixelmatchOptions: {}
+            },
             screenshotName: 'not_ever',
             generateDiff: 'never'
           }
@@ -127,15 +137,18 @@ describe('plugin', () => {
           expect(result.images.diff).toBeUndefined()
           expect(result.baseGenerated).toBeUndefined()
           expect(result.error).toBeDefined()
-          const isDiffGenerated = existsSync(path.join(options.diffDirectory as string, 'not_ever.png'))
+          const isDiffGenerated = existsSync(path.join(options.diffDirectory, 'not_ever.png'))
           expect(isDiffGenerated).toBe(false)
         })
         it('should not generate a diff image if threshold is higher than mismatched percentage ', async () => {
-          const options: CompareSnapshotOptions = {
+          const options: VisualRegressionOptions = {
             ...baseCompareOptions,
             screenshotAbsolutePath: path.join('./src/fixtures/assets/mod/not_ever.png'),
-            specName: '',
-            errorThreshold: 1,
+            pluginOptions: {
+              errorThreshold: 1,
+              failSilently: false,
+              pixelmatchOptions: {}
+            },
             screenshotName: 'not_ever'
           }
           const result = await compareSnapshots(options)
@@ -146,16 +159,19 @@ describe('plugin', () => {
           expect(result.images.diff).toBeUndefined()
           expect(result.baseGenerated).toBeUndefined()
           expect(result.error).toBeUndefined()
-          const isDiffGenerated = existsSync(path.join(options.diffDirectory as string, 'not_ever.png'))
+          const isDiffGenerated = existsSync(path.join(options.diffDirectory, 'not_ever.png'))
           expect(isDiffGenerated).toBe(false)
         })
 
         it('should generate a diff image if generateDiff is set to always', async () => {
-          const options: CompareSnapshotOptions = {
+          const options: VisualRegressionOptions = {
             ...baseCompareOptions,
-            specName: '',
             screenshotAbsolutePath: path.join('./src/fixtures/assets/base/enjuto.png'),
-            errorThreshold: 0,
+            pluginOptions: {
+              errorThreshold: 0,
+              failSilently: false,
+              pixelmatchOptions: {}
+            },
             generateDiff: 'always'
           }
           const result = await compareSnapshots(options)
@@ -166,7 +182,7 @@ describe('plugin', () => {
           expect(result.images.diff).toBeDefined()
           expect(result.baseGenerated).toBeUndefined()
           expect(result.error).toBeUndefined()
-          const isDiffGenerated = existsSync(path.join(options.diffDirectory as string, 'enjuto.png'))
+          const isDiffGenerated = existsSync(path.join(options.diffDirectory, 'enjuto.png'))
           expect(isDiffGenerated).toBe(true)
           rmSync('./src/fixtures/diff', { recursive: true })
         })