+++ tests/integration/cypress/support/e2e.ts
@@ -1,2 +1,9 @@
 import { addCompareSnapshotCommand } from 'cypress-visual-regression/dist/command'
-addCompareSnapshotCommand()
+addCompareSnapshotCommand({
+  pixelmatchOptions: {
+    threshold: 0.1
+  },
+  errorThreshold: 0,
+  failSilently: false,
+  capture: 'fullPage'
+})